requirements/tests.txt
-r common.txt

psycopg2==2.6.1
model-mommy==1.2.6
coverage==4.0.3
py==1.4.31
pytest==2.8.5
pytest-cov==2.2.0
pytest-cover==3.0.0
pytest-django==2.9.1
